% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesClint_runthru.R
\name{BayesClint_postprocess}
\alias{BayesClint_postprocess}
\title{Post-process the posterior sampling results}
\usage{
BayesClint_postprocess(zeta2mcmc, kappa2mcmc, mu2mcmc = NULL, C, K)
}
\arguments{
\item{zeta2mcmc}{MCMC samples of the cell type labels}

\item{kappa2mcmc}{MCMC samples of the spatial domain labels}

\item{mu2mcmc}{MCMC samples of the mu parameter}

\item{C}{number of cell types}

\item{K}{number of spatial domains}
}
\value{
Returns the estimated cell type and spatial domain cluster labels, that have been corrected for label switching.
Also returns the MCMC draws for mu that have been corrected for cell type label switching (if the original mu2mcmc is provided).
Also returns an estimate of the cell type compositions theta calculated on the basis of the corrected cluster labels.
}
\description{
Post-process the posterior sampling results to address the label switching issue associated with
the sampling of cell type labels and spatial domain labels based on the ECR-1 algorithm, estimate
the cell type labels and spatial domain labels as the mode of all their posterior samples, and
estimate the cell type composition in each spatial domain based on the final estimates of cell type and
spatial domain labels.
}
\examples{
# See documentation for BayesClint_run
}
