% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesClint_runthru.R
\name{BayesClint_preprocess}
\alias{BayesClint_preprocess}
\title{Pre-process gene expression data}
\usage{
BayesClint_preprocess(
  cnts,
  info,
  K,
  cutoff_sample = 100,
  cutoff_feature = 0.1,
  doBatchCorrect = F
)
}
\arguments{
\item{cnts}{list of gene count matrices across tissue sections, of dimensions number of genes by number of cells. Make sure both the rows and columns are labelled with gene names and cell names.}

\item{info}{list of spatial coordinates across tissue sections, of dimensions number of cells by number of coordinates.}

\item{K}{number of spatial domains (for Potts model calculations)}

\item{cutoff_sample}{only retain quality cells with greater than cutoff_sample counts across genes}

\item{cutoff_feature}{only retain quality genes with at least cutoff_feature percent of cells with nonzero counts}

\item{doBatchCorrect}{Logical. Whether to perform batch effect adjustment with Seurat v3 (Stuart et al., 2019) to align expression data from different tissue sections.}
}
\description{
A function to pre-process the data to get it ready for the BayesClint algorithm.
}
\examples{
BayesClint_preprocess()
}
