% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesClint_runthru.R
\name{BayesClint_run}
\alias{BayesClint_run}
\title{Running the BayesClint MCMC algorithm}
\usage{
BayesClint_run(
  dataList,
  info,
  C,
  K,
  r,
  nbrsample,
  burnin,
  probvarsel,
  nc_quant_list = NULL,
  chainNbr = 1
)
}
\arguments{
\item{dataList}{pre-processed dataset, like the one outputted by BayesClint_preprocess}

\item{info}{list of spatial coordinates across tissue sections, of dimensions number of cells by number of coordinates. This should be the info returned by the BayesClint_preprocess's quality control.}

\item{C}{number of cell types}

\item{K}{number of spatial domains}

\item{r}{number of components in the factor model}

\item{nbrsample}{number of MCMC iterations retained}

\item{burnin}{number of initial MCMC iterations discarded as burn-in}

\item{probvarsel}{prior probability for an entry in the factor loadings matrix to be nonzero}

\item{nc_quant_list}{the quantities used to compute the normalizing constant of the Potts model within the MCMC.
The nc_quant_list object will already be calculated within the BayesClint_run function;
however, since this step takes a long time, one can import nc_quant_list (precalculated by an earlier run of BayesClint_run)
through this argument to skip this computationally intensive step. However, if the nc_quant_list argument is set to NULL,
BayesClint_run will compute it and return it as one of the outputs.}

\item{chainNbr}{MCMC chain number. The default is 1 for one MCMC number 1. If you want to run N multiple MCMC chains, make a loop in R as for(i in 1:N) BayesClint_run(..., chainNbr=i)}
}
\value{
the MCMC results from the BayesClint algorithm, and optionally the intermediary data file nc_quant_list.
}
\description{
A function to pre-process the data to get it ready for the BayesClint algorithm.
}
\examples{
BayesClint_run()
}
