% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesClint_runthru.R
\name{simulate}
\alias{simulate}
\title{Generate datasets as featured in BayesClint manuscript}
\usage{
simulate(f_idx = 9)
}
\arguments{
\item{f_idx}{The index of the row of the simulation results tables (e.g. Table 2) in the manuscript corresponding to the desired scenario. For example, f_idx = 9 corresponds to the scenario with N = 3, irregular theta composition, total number of genes = 200, and number of differentiating genes = 40.}
}
\value{
Returns a dataset generated according to one of the scenarios in the manuscript, in the form of a list. Below are the elements of the list:
\item{simver_expr}{a list of matrices (dimension number of genes by number of cells) of the natural log of the true Poisson rates for each gene and each cell, for all tissue samples.}
\item{simver_info}{a list of matrices (dimension number of cells by 4) of the x- and y-coordinates, and the true cell-type and spatial-domain labels, for all tissue samples.}
\item{U}{a list of matrices (dimension number of cells by number of factors) of the true factors corresponding to each cell, for all tissue samples.}
\item{Xcounts}{a list of matrices (dimension number of cells by number of genes) of the observed raw count for each gene and each cell, for all tissue samples. These are the data that the BayesClint algorithm will directly analyze, along with the x- and y- coordinates of the cells.}
}
\description{
This function allows the user to generate datasets from any of the scenarios included in the manuscript
}
\examples{
test_dataset1 <- simulate(f_idx = 1)
cnts <- lapply(test_dataset1$Xcounts, t)
info <- test_dataset1$simver_info
preprocess_results <- BayesClint_preprocess(cnts, info, cutoff_sample = 0, cutoff_feature = 0,
                                            doBatchCorrect = T, doScaling = T)
set.seed(1057)
# The below command takes about 15 minutes to run on the above simulated dataset.
run_results_sim <- BayesClint_run(dataList = preprocess_results$dataList, info = preprocess_results$info,
                                  C = 4, K = 4, r = 4,
                                  nbrsample = 15000,
                                  burnin = 6500,
                                  probvarsel = 0.05,
                                  nc_quant_list = NULL,
                                  chainNbr = 1)
postprocess_results <- BayesClint_postprocess(run_results_sim$zeta2mcmc,
                                              run_results_sim$kappa2mcmc,
                                              run_results_sim$mu2mcmc,
                                              C = 4, K = 4)
# test the ARI performances
true_zeta <- do.call("c", lapply(preprocess_results$info, function(section) section$zeta))
mclust::adjustedRandIndex(true_zeta, do.call("c", postprocess_results$zeta_est))
true_kappa <- do.call("c", lapply(preprocess_results$info, function(section) section$kappa))
mclust::adjustedRandIndex(true_kappa, do.call("c", postprocess_results$kappa_est))
}
\references{
Sheng, A., Chekouo, T., Safo, S. E. (2026). BayesClint: Bayesian multi-scale clustering and multi-sample integration with feature selection for spatial transcriptomics data.
}
