% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesClint_runthru.R
\name{BayesClint_run}
\alias{BayesClint_run}
\title{Running the BayesClint MCMC algorithm}
\usage{
BayesClint_run(
  dataList,
  info,
  C,
  K,
  r,
  nbrsample,
  burnin,
  probvarsel,
  nc_quant_list = NULL,
  chainNbr = 1
)
}
\arguments{
\item{dataList}{pre-processed dataset, like the one outputted by BayesClint_preprocess}

\item{info}{list of spatial coordinates across tissue samples, of dimensions number of cells by number of coordinates. This should be the info returned by the BayesClint_preprocess's quality control.}

\item{C}{number of cell types}

\item{K}{number of spatial domains}

\item{r}{number of components in the factor model}

\item{nbrsample}{number of MCMC iterations retained}

\item{burnin}{number of initial MCMC iterations discarded as burn-in}

\item{probvarsel}{prior probability for an entry in the factor loadings matrix to be nonzero}

\item{nc_quant_list}{the quantities used to compute the normalizing constant of the Potts model within the MCMC.
The nc_quant_list object will already be calculated within the BayesClint_run function;
however, since this step takes a long time, one can import nc_quant_list (pre-calculated by an earlier run of BayesClint_run)
through this argument to skip this computationally intensive step. However, if the nc_quant_list argument is set to NULL,
BayesClint_run will compute it and return it as one of the outputs.}

\item{chainNbr}{MCMC chain number. The default is 1 for one MCMC number 1. If you want to run N multiple MCMC chains, make a loop in R as for(i in 1:N) BayesClint_run(..., chainNbr=i)}
}
\value{
Returns the list of MCMC results from the BayesClint algorithm, and optionally the intermediary data file nc_quant_list. The elements of the list are described below.
\item{VarSelMean}{A row-vectorized, P (number of genes) by r matrix of the posterior probabilities of inclusion for each gene, within each component.}
\item{VarSelMeanGlobal}{A length-P vector of the posterior probabilities of inclusion for the genes across components.}
\item{zeta2mcmc}{MCMC samples of the cell type labels zeta}
\item{kappa2mcmc}{MCMC samples of the spatial domain labels kappa}
\item{mu2mcmc}{MCMC samples of the cell type means mu}
\item{theta2mcmc}{MCMC samples of the cell type compositions for each spatial domain, theta}
\item{beta2mcmc}{Potts parameter beta for each tissue sample}
}
\description{
This function performs the Markov Chain Monte Carlo sampling for the BayesClint model.
As shown in Examples, the BayesClint pipeline begins with inputting the data into BayesClint_preprocess;
then, the pre-processed data is inputted into BayesClint_run; finally, the MCMC output from BayesClint_run
is inputted into BayesClint_postprocess to get the final estimated cell type and spatial domain cluster labels,
as well as the estimated cell type compositions of each spatial domain.
}
\examples{
data("starmap_mpfc_subset")
cnts <- starmap_mpfc_subset$starmap_cnts
info <- starmap_mpfc_subset$starmap_info
preprocess_results <- BayesClint_preprocess(cnts, info, cutoff_sample = 100, cutoff_feature = 0.1,
doBatchCorrect = T, doScaling = T, k.filter = 50)
set.seed(1057)
run_results <- BayesClint_run(dataList = preprocess_results$dataList, info = preprocess_results$info,
C = 15, K = 4, r = 9,
nbrsample = 15000,
burnin = 6500,
probvarsel = 0.05,
nc_quant_list = NULL,
chainNbr = 1)
postprocess_results <- BayesClint_postprocess(run_results$zeta2mcmc,
run_results$kappa2mcmc,
run_results$mu2mcmc,
C = 15, K = 4)
# test the ARI performances (beware that this is hard-coded for the mPFC dataset)
true_zeta <- do.call("c", lapply(preprocess_results$info, function(section) section$c))
mclust::adjustedRandIndex(true_zeta, do.call("c", postprocess_results$zeta_est))
true_kappa <- do.call("c", lapply(preprocess_results$info, function(section) section$z))
mclust::adjustedRandIndex(true_kappa, do.call("c", postprocess_results$kappa_est))
}
\references{
Sheng, A., Chekouo, T., Safo, S. E. (2026). BayesClint: Bayesian multi-scale clustering and multi-sample integration with feature selection for spatial transcriptomics data.
}
